+++ src/AutoMoq/AutoMoq.csproj
@@ -75,9 +75,9 @@
     <Compile Include="Helpers\AutoMoqTestFixture.cs" />
     <Compile Include="Helpers\with_automoqer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Unity\AutoMockingBuilderStrategy.cs" />
+    <Compile Include="Unity\MoqBuilderStrategy.cs" />
     <Compile Include="Unity\AutoMockingContainerExtension.cs" />
-    <Compile Include="Unity\Base.cs" />
+    <Compile Include="Unity\AutoMockingBuilderStrategy.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="License.txt" />
