+++ src/AutoMoq/Unity/AutoMockingBuilderStrategy.cs
@@ -1,37 +1,103 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
+using Microsoft.Practices.ObjectBuilder2;
 using Microsoft.Practices.Unity;
-using Moq;
 
 namespace AutoMoq.Unity
 {
-    internal class AutoMockingBuilderStrategy : Base
+    internal abstract class AutoMockingBuilderStrategy : BuilderStrategy
     {
-        private readonly MockRepository mockRepository;
+        private readonly IUnityContainer container;
+        private readonly IEnumerable<Type> registeredTypes;
 
         public AutoMockingBuilderStrategy(IEnumerable<Type> registeredTypes, IUnityContainer container)
-            : base(registeredTypes, container)
         {
-            mockRepository = new MockRepository(MockBehavior.Loose);
+            this.registeredTypes = registeredTypes;
+            this.container = container;
         }
 
-        internal override MethodInfo GenerateAnInterfaceMockCreationMethod(Type type)
+        public override void PreBuildUp(IBuilderContext context)
         {
-            var createMethodWithNoParameters = mockRepository.GetType().GetMethod("Create", EmptyArgumentList());
-            return createMethodWithNoParameters.MakeGenericMethod(type);
+            var type = GetTheTypeFromTheBuilderContext(context);
+            if (AMockObjectShouldBeCreatedForThisType(type))
+            {
+                var mock = CreateAMockTrackedByAutoMoq(type);
+                context.Existing = mock.Object;
+            }
+
+            if (type.GetConstructors().Any() == false) return;
+
+            LoadAbstractDependenciesInTheGreediestConstructor(type);
         }
 
-        internal override dynamic CreateAMockObject(Type type)
+        internal void LoadAbstractDependenciesInTheGreediestConstructor(Type type)
         {
-            var createMethod = GenerateAnInterfaceMockCreationMethod(type);
+            var constructor = type.GetConstructors().OrderByDescending(x => x.GetParameters().Count()).First();
+            var abstractParameters = constructor.GetParameters()
+                .Where(x => x.ParameterType.IsAbstract)
+                .Where(x => x.ParameterType.IsInterface == false)
+                .Select(x => x.ParameterType);
 
-            return InvokeTheMockCreationMethod(createMethod);
+            foreach (var abstractParameter in abstractParameters)
+            {
+                var mock = CreateAMockTrackedByAutoMoq(abstractParameter);
+                try
+                {
+                    container.Resolve(abstractParameter);
+                }
+                catch
+                {
+                    var mockObject = mock.Object as object;
+                    container.RegisterInstance(abstractParameter, mockObject);
+                }
+            }
         }
 
-        private dynamic InvokeTheMockCreationMethod(MethodInfo createMethod)
+        internal dynamic CreateAMockTrackedByAutoMoq(Type type)
         {
-            return (Mock) createMethod.Invoke(mockRepository, new object[] {new List<object>().ToArray()});
+            var mock = CreateAMockObject(type);
+            var autoMoqer = container.Resolve<AutoMoqer>();
+            autoMoqer.SetMock(type, mock);
+            return mock;
+        }
+
+        internal bool AMockObjectShouldBeCreatedForThisType(Type type)
+        {
+            return ThisTypeIsNotAFunction(type) &&
+                   ThisTypeIsNotRegistered(type) &&
+                   ThisIsNotTheTypeThatIsBeingResolvedForTesting(type);
+        }
+
+        internal bool ThisIsNotTheTypeThatIsBeingResolvedForTesting(Type type)
+        {
+            var mocker = container.Resolve<AutoMoqer>();
+            return (mocker.ResolveType == null || mocker.ResolveType != type);
+        }
+
+        internal static bool ThisTypeIsNotAFunction(Type type)
+        {
+            return type.Name != "Func`1";
+        }
+
+        internal static Type GetTheTypeFromTheBuilderContext(IBuilderContext context)
+        {
+            return (context.OriginalBuildKey).Type;
+        }
+
+        internal bool ThisTypeIsNotRegistered(Type type)
+        {
+            return registeredTypes.Any(x => x.Equals(type)) == false;
+        }
+
+        internal abstract dynamic CreateAMockObject(Type type);
+
+        internal abstract MethodInfo GenerateAnInterfaceMockCreationMethod(Type type);
+
+        internal static Type[] EmptyArgumentList()
+        {
+            return new[] {typeof (object[])};
         }
     }
 }
